chore: wire up ofrep server
