+++ internal/server/evaluation/ofrep/server.go
@@ -44,3 +44,23 @@ func (s *Server) RegisterGRPC(server *grpc.Server) {
 func (s *Server) SkipsAuthorization(ctx context.Context) bool {
 	return true
 }
+
+func (s *Server) EvaluateFlag(ctx context.Context, r *ofrep.EvaluateFlagRequest) (*ofrep.EvaluationResponse, error) {
+	return s.bridge.OFREPFlagEvaluation(ctx, r)
+}
+
+func (s *Server) EvaluateBulk(ctx context.Context, r *ofrep.EvaluateBulkRequest) (*ofrep.BulkEvaluationResponse, error) {
+	return s.bridge.OFREPFlagEvaluationBulk(ctx, r)
+}
+
+// GetProviderConfiguration returns the configuration set by the running flipt instance.
+func (s *Server) GetProviderConfiguration(_ context.Context, _ *ofrep.GetProviderConfigurationRequest) (*ofrep.GetProviderConfigurationResponse, error) {
+	return &ofrep.GetProviderConfigurationResponse{
+		Name: "flipt",
+		Capabilities: &ofrep.Capabilities{
+			FlagEvaluation: &ofrep.FlagEvaluation{
+				SupportedTypes: []string{"string", "boolean"},
+			},
+		},
+	}, nil
+}
